Fix life

+++ python/kiss_icp/pipeline.py
@@ -221,11 +221,7 @@ class OdometryPipeline:
         results_dir = os.path.join(os.path.realpath(out_dir), get_current_timestamp())
         latest_dir = os.path.join(os.path.realpath(out_dir), "latest")
         os.makedirs(results_dir, exist_ok=True)
-        (
-            os.unlink(latest_dir)
-            if os.path.exists(latest_dir) or os.path.islink(latest_dir)
-            else None
-        )
+        os.unlink(latest_dir) if os.path.exists(latest_dir) or os.path.islink(latest_dir) else None
         os.symlink(results_dir, latest_dir)
         return results_dir
